feat: extend `to_layout` options (#2763)
* refactor: make `to_layout` support finer granularity

* feat: extend customisation of `to_layout`

* fix: pass `allow_record` to `from_iter`

* fix: add `allow_none`

* fix: type promotion

* fix: always return input for non-promoted scalars

* refactor: remove useless `to_layout`

* refactor: add `string_policy`

* fix: promote strings

* test: basic strings

* test: more tests

* test: check against default ArrayBuilder type

* refactor: rename `allow_other` → `allow_unknown`

* fix: more rename fixes

* test: explicitly check string coercion for with_field

+++ src/awkward/_nplikes/typetracer.py
@@ -757,7 +757,7 @@ class TypeTracer(NumpyLike):
         else:
             length = unknown_length
 
-        default_int_type = np.int64 if (ak._util.win or ak._util.bits32) else np.int32
+        default_int_type = np.int32 if (ak._util.win or ak._util.bits32) else np.int64
         return TypeTracerArray._new(dtype or default_int_type, (length,))
 
     def meshgrid(
